feat: 오브제 생성/수정 코드 통합
closes #145
